+++ lib/plugins/index.js
@@ -19,7 +19,7 @@ module.exports = [
   require('./plugin/install.js'),
   require('./plugin/uninstall/uninstall.js'),
   require('./plugin/list.js'),
-  require('./plugin/search/search.js'),
+  require('./plugin/search.js'),
   require('./aws/configCredentials.js'),
   require('./aws/provider.js'),
   require('./aws/common/index.js'),
