+++ lib/plugins/plugin/search.js
@@ -2,7 +2,7 @@
 
 const BbPromise = require('bluebird');
 const chalk = require('chalk');
-const pluginUtils = require('../lib/utils');
+const pluginUtils = require('./lib/utils');
 
 class PluginSearch {
   constructor(serverless, options) {
